mode-watermarking/requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0

# Numerical and image processing
numpy>=1.21.0
pillow>=9.0.0
scipy>=1.9.0

# Image quality metrics
lpips>=0.1.4
scikit-image>=0.19.0
pytorch-fid>=0.3.0

# Machine learning utilities
scikit-learn>=1.2.0

# Visualization (optional, but needed for evaluation)
matplotlib>=3.5.0
seaborn>=0.12.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0

# Logging and experiment tracking (optional)
wandb>=0.15.0
tensorboard>=2.13.0

# Data download (optional, for COCO dataset)
kagglehub>=0.2.0
